requirements.txt
click==8.0.1
loguru==0.5.3
huey>=2
python-dotenv==0.19.0
greenlet==1.1.0
gevent